implemented authentication

+++ src/main/java/com/gft/delivery/dto/EstoqueDto.java
@@ -1,5 +1,7 @@
 package com.gft.delivery.dto;
 
+import java.math.BigDecimal;
+
 import org.springframework.hateoas.RepresentationModel;
 import org.springframework.hateoas.server.core.Relation;
 
@@ -18,29 +20,37 @@ public class EstoqueDto extends RepresentationModel<EstoqueDto> {
 	private Long id;
 	private Produto produto;
 	private int quantity;
-	private Double sellingPrice;
+	private BigDecimal sellingPrice;
+	
 	public Long getId() {
 		return id;
 	}
+	
 	public void setId(Long id) {
 		this.id = id;
 	}
+	
 	public Produto getProduto() {
 		return produto;
 	}
+	
 	public void setProduto(Produto produto) {
 		this.produto = produto;
 	}
+	
 	public int getQuantity() {
 		return quantity;
 	}
+	
 	public void setQuantity(int quantity) {
 		this.quantity = quantity;
 	}
-	public Double getSellingPrice() {
+	
+	public BigDecimal getSellingPrice() {
 		return sellingPrice;
 	}
-	public void setSellingPrice(Double sellingPrice) {
+	
+	public void setSellingPrice(BigDecimal sellingPrice) {
 		this.sellingPrice = sellingPrice;
 	}
